[step3] refactor: getLotteryResult 정렬리스트 제거하고 보너스, 비 보너스 점수 나누어서 map으로 가져오도록 수정

+++ src/test/java/lottery/domain/LotteryResultTest.java
@@ -55,7 +55,7 @@ class LotteryResultTest {
     @DisplayName("getTotalCashPrize 테스트")
     public void getTotalCashPrize() {
         // given
-        long count = 10;
+        int count = 10;
         long expected = LotteryResult.FIVE_MATCHES.toDto(0)
                 .getCashPrize()
                 .multiply(count);
@@ -71,7 +71,7 @@ class LotteryResultTest {
     @DisplayName("toDto 테스트")
     public void toDto() {
         // given
-        long count = 10;
+        int count = 10;
         LotteryResultDto expected = LotteryResult.FIVE_MATCHES.toDto(count);
 
         // when
